third_party/WebKit/LayoutTests/dom/domparsing/xmlserializer-attribute-special-namespaces-expected.txt

Tests the serialization of special XML namespaces on attributes, as reported in bug 395950.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xmlDocument.querySelector("inner1").getAttributeNS("http://www.w3.org/1999/xlink", "href") is "http://www.google.com"
PASS xmlDocument.querySelector("inner1").getAttributeNS("http://www.w3.org/XML/1998/namespace", "attr") is "value"
PASS xmlDocument.querySelector("inner2").getAttributeNS("http://www.w3.org/1999/xlink", "href") is "http://www.google.com"
PASS xmlDocument.querySelector("inner2").getAttribute("xl:href") is "http://www.google.com"
PASS xmlDocument.querySelector("inner2").getAttributeNS("http://www.w3.org/XML/1998/namespace", "attr") is "value"
PASS xmlDocument.querySelector("inner2").getAttribute("xml:attr") is "value"
PASS parsedDoc.querySelector("inner1").getAttributeNS("http://www.w3.org/1999/xlink", "href") is "http://www.google.com"
PASS parsedDoc.querySelector("inner1").getAttributeNS("http://www.w3.org/XML/1998/namespace", "attr") is "value"
PASS parsedDoc.querySelector("inner2").getAttributeNS("http://www.w3.org/1999/xlink", "href") is "http://www.google.com"
PASS parsedDoc.querySelector("inner2").getAttribute("xl:href") is "http://www.google.com"
PASS parsedDoc.querySelector("inner2").getAttributeNS("http://www.w3.org/XML/1998/namespace", "attr") is "value"
PASS parsedDoc.querySelector("inner2").getAttribute("xml:attr") is "value"
PASS xmlString.indexOf("xmlns:xml") is -1
PASS xmlString.indexOf("xmlns:xlink") is not -1
PASS xmlString.indexOf("xmlns:xl") is not -1
PASS successfullyParsed is true

TEST COMPLETE
